Share logo mark loading
