+++ LongevityWorldCup.Website/Business/AthleteOgImageService.cs
@@ -290,7 +290,7 @@ public sealed class AthleteOgImageService
 
         try
         {
-            using var logo = await LoadLogoMarkAsync();
+            using var logo = await ImageLogo.LoadMarkAsync(_logoPath);
             using var smallLogo = logo.Clone(ctx => ctx.Resize(new ResizeOptions
             {
                 Size = new Size(54, 54),
@@ -322,33 +322,6 @@ public sealed class AthleteOgImageService
                 VerticalAlignment = VerticalAlignment.Top
             }, "LONGEVITY\nWORLD CUP", NameColor);
         });
-    }
-
-    private async Task<Image<Rgba32>> LoadLogoMarkAsync()
-    {
-        await using var logoStream = File.OpenRead(_logoPath);
-        var logo = await Image.LoadAsync<Rgba32>(logoStream);
-        logo.ProcessPixelRows(accessor =>
-        {
-            for (var y = 0; y < accessor.Height; y++)
-            {
-                var row = accessor.GetRowSpan(y);
-                for (var x = 0; x < row.Length; x++)
-                {
-                    var pixel = row[x];
-                    var brightness = (pixel.R + pixel.G + pixel.B) / 3f;
-                    if (brightness < 110f)
-                    {
-                        row[x] = Color.Transparent;
-                        continue;
-                    }
-
-                    var alpha = (byte)Math.Clamp((brightness - 110f) * 2.4f, 0f, pixel.A);
-                    row[x] = new Rgba32(255, 255, 255, alpha);
-                }
-            }
-        });
-        return logo;
     }
 
     private static void DrawScoreboardMetricRow(
